sqlite_to_postgres/pyproject.toml
[tool.poetry]
name = "sqlite-to-postgres"
version = "0.1.0"
description = ""
authors = ["pakum <makcayka2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
